backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace
diffusers>=0.27.0
transformers>=4.41.0
accelerate>=0.31.0
safetensors>=0.4.3

# Image & video
Pillow>=10.0.0
imageio[ffmpeg]>=2.34.0

# Utilities
numpy>=1.26.0

# Optional (ZIP, etc.)
python-multipart>=0.0.9

fastapi
uvicorn
pillow
torch
diffusers
transformers
opencv-python-headless
